+++ .travis.yml
@@ -10,7 +10,7 @@ script:
     - go build
     - go test -v ./ast/util
     - go test -v ./vm
-    - go test -v ./builtins
+    - go test -v ./core
     - go test -v ./packages
     - ./t/test.sh
     - ./anko ./_example/scripts/term.ank
