+++ core/core_test.go
